+++ src/Controllers/UserController.php
@@ -5,6 +5,7 @@ namespace MarkVilludo\Permission\Controllers;
 use MarkVilludo\Permission\Models\Permission;
 use App\Http\Controllers\Controller;
 use MarkVilludo\Permission\Models\Role;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Http\Request;
 use App\User;
 use Session;
@@ -25,14 +26,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::all();
-
         if (View::exists('users.index')) {
-            return view('users.index')->with('users', $users);
+            return view('users.index');
         } else if (View::exists('superadmin.users.index')) {
-            return view('superadmin.users.index')->with('users', $users);
-        } else  {
-            return view('laravel-permission::users.index')->with('users', $users);
+            return view('superadmin.users.index');
+        } else {
+            return view('laravel-permission::users.index');
         }
     }
 
@@ -44,16 +43,20 @@ class UserController extends Controller
     public function create()
     {
         $roles = Role::get();
+        $permissions = Permission::all();
+        $data['roles'] = $roles;
+        $data['permissions'] = $permissions;
 
         if (View::exists('users.create')) {
-            return view('users.create', ['roles'=>$roles]);
+            return view('users.create', $data);
         } else if (View::exists('superadmin.users.create')) {
-            return view('superadmin.users.create', ['roles'=>$roles]);
+            return view('superadmin.users.create', $data);
         } else {
-            return view('laravel-permission::users.create', ['roles'=>$roles]);
+            return view('laravel-permission::users.create', $data);
         }
     }
 
+    /**
     /**
      * Store a newly created resource in storage.
      *
@@ -67,11 +70,23 @@ class UserController extends Controller
         $this->validate($request, [
             'first_name'=>'required|max:120',
             'last_name'=>'required|max:120',
-            'email'=>'required|email|unique:users',
-            'password'=>'required|min:6|confirmed'
+            'email'=>'required|email|unique:users'
         ]);
 
-        $user = User::create($request->only('email', 'first_name', 'last_name', 'password'));
+        $user = new User;
+        $randomPassword = str_random(8);
+
+        $user->email = $request->email;
+        $user->first_name = $request->first_name;
+        $user->last_name = $request->last_name;
+        $user->type = 'admin';
+        $user->password = bcrypt($randomPassword);
+        $user->is_expire_access = $request->is_expire_access;
+        $user->expiration_date = $request->expiration_date;
+        $user->save();
+
+
+        $this->sendCredentials($user, $randomPassword);
 
         $roles = $request['roles'];
 
@@ -92,6 +107,16 @@ class UserController extends Controller
         } else {
             return view('laravel-permission::users.index')->with('users', $users)->with('flash_message','User successfully added.');
         }
+    }
+
+    public function sendCredentials($user, $randomPassword)
+    {
+        $emails = $user->email;
+        Mail::send('emails.send_credentials', ['project_title' => 'MPBL', 'first_name' => $user->first_name, 'last_name' => $user->last_name, 'email' => $user->email, 'newPassword' => $randomPassword], function ($message) use ($emails) {
+           $message->from(env('MAIL_FROM', 'mark.villudo@synergy88digital.com'));
+           $message->to($emails);
+           $message->subject("MPBL Credentials");
+       });
     }
 
     /**
@@ -139,13 +164,19 @@ class UserController extends Controller
         $this->validate($request, [
             'first_name'=>'required|max:120',
             'last_name'=>'required|max:120',
-            'email'=>'required|email|unique:users,email,'.$id,
-            'password'=>'required|min:6|confirmed'
+            'email'=>'required|email|unique:users,email,'.$id
         ]);
 
-        $input = $request->only(['name', 'email', 'password']);
+        $user->email = $request->email;
+        $user->first_name = $request->first_name;
+        $user->last_name = $request->last_name;
+        $user->type = 'admin';
+        $user->is_expire_access = $request->is_expire_access;
+        $user->expiration_date = $request->expiration_date;
+        $user->update();
+
         $roles = $request['roles'];
-        $user->fill($input)->save();
+     
 
         //collect roles name and syn in user roles
         $rolesArray = [];
@@ -181,17 +212,18 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
+        $user = User::find($id);
 
-        if (View::exists('users.index')) {
-            return redirect()->route('users.index')->with('flash_message', 'User successfully deleted.');
-        } else if (View::exists('superadmin.users.index')) {
-            return redirect()->route('superadmin.users.index')->with('flash_message', 'User successfully deleted.');
+        if ($user) {
+            $user->delete();
+            $data['message'] = 'Successfully deleted user.';
+            $statusCode = 200;
+
         } else {
-            return redirect()->route('laravel-permission::users.index')
-                ->with('flash_message',
-                 'User successfully deleted.');
+            $data['message'] = 'Not found user.';
+            $statusCode = 404;
         }
+
+        return response()->json($data, $statusCode);
     }
 }
